Updated oneOfK function to allow class labels in any range, e.g. [0,1], 1:K, [-1,+1] or even {'label1',label2',lable3'}.
Modified LogregDist class to keep track of the original support of y and pass this to the predictive distributions. 

Created a SampleDistDiscrete class to handle cases where rows are not d-dimensional samples but rather probabilities over a discrete support as in the case with LogregDist.predict with 'mc'.

Updated LogregDist examples to make use of the new class support management, i.e. there is now no need to convert class labels to 1:K before calling LogregDist.fit, [0,1] etc work just fine.